pyproject.toml
[project]
name = "gis-assessor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cyclopts>=3.20.0",
    "dspy-ai>=2.6.27",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.402",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]
